pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "athanor"
version = "0.1.0"
description = "Domain-agnostic automated science infrastructure"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.40.0",
    "arxiv>=2.1.0",
    "networkx>=3.3",
    "sentence-transformers>=3.0.0",
    "pydantic>=2.0.0",
    "numpy>=1.26.0",
    "scipy>=1.13.0",
    "pypdf>=4.0.0",
    "requests>=2.31.0",
    "pyyaml>=6.0.1",
    "click>=8.1.0",
    "pyvis>=0.3.2",
    "plotly>=5.23.0",
    "python-dotenv>=1.0.0",
    "rich>=13.7.0",
    "tqdm>=4.66.0",
    "scikit-learn>=1.5.0",
]

[project.scripts]
athanor = "athanor.cli:cli"

[tool.setuptools.packages.find]
where = ["."]
include = ["athanor*"]
